% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{run.gui}
\alias{run.gui}
\title{Run the simulation in a graphical interface}
\usage{
run.gui(map.filename = system.file("extdata", "map.csv", package =
  "li19trustmodel"))
}
\description{
Simulate the trust model and the mobile network and iterate
}
\keyword{model}
\keyword{run}
\keyword{simulate}
\keyword{simulation}
\keyword{trust}
