% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseStation.R
\docType{class}
\name{BaseStation-class}
\alias{BaseStation-class}
\alias{BaseStation}
\title{A Base Station Class}
\description{
This class defines the base stations within the trust model, these are
essentially gateways.
}

\examples{
BaseStation(1, 1)
}
\keyword{Base}
\keyword{Gateway}
\keyword{Station}
