% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Field.R
\docType{class}
\name{Field-class}
\alias{Field-class}
\alias{Field}
\title{A Field Class}
\description{
This class is a field containing the base stations and their signals
}
\section{Methods}{

\describe{
\item{\code{get.tile(location)}}{Get the tile at the location if there is one, otherwise NA}

\item{\code{size()}}{Get the size of the field}
}}

\examples{
Field(c(BaseStation(1, 1), BaseStation(2, 2)))
}
\keyword{Base}
\keyword{Map}
\keyword{Station}
