% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{run.simulation}
\alias{run.simulation}
\title{Run the simulation in a console interface}
\usage{
run.simulation(total.time, map.filename = system.file("extdata",
  "map.csv", package = "li19trustmodel"))
}
\description{
Simulate the trust model and the mobile network and iterate for an amount
of time
}
\keyword{model}
\keyword{run}
\keyword{simulate}
\keyword{simulation}
\keyword{trust}
